debugging code

+++ src/main/java/com/ict/careus/dto/response/JwtResponse.java
@@ -8,12 +8,12 @@ import java.util.List;
 public class JwtResponse {
     private String token;
     private String type = "Bearer";
-    private int id;
+    private long id;
     private String username;
     private String phoneNumber;
     private List<String> roles;
 
-    public JwtResponse(String accessToken, int id, String username, String phoneNumber, List<String> roles) {
+    public JwtResponse(String accessToken, long id, String username, String phoneNumber, List<String> roles) {
         this.token = accessToken;
         this.id = id;
         this.username = username;
